Move GraphVerifier to `dag` module
