+++ test/unit/optimizer/test_gp_operators.py
@@ -7,7 +7,7 @@ from fedot.core.composer.advisor import PipelineChangeAdvisor
 from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
 from fedot.core.optimisers.objective.data_objective_builder import DataObjectiveBuilder
 from fedot.core.composer.gp_composer.specific_operators import boosting_mutation
-from fedot.core.dag.validation_rules import DEFAULT_DAG_RULES
+from fedot.core.dag.verification_rules import DEFAULT_DAG_RULES
 from fedot.core.data.data import InputData
 from fedot.core.log import default_log
 from fedot.core.optimisers.adapters import DirectAdapter, PipelineAdapter
